More intelligent alsa_compare_clients function, with printf's for testing.
The lists are sorted alphabetically, so the order in which clients are started,
doesn't matter.

If there are more clients than before, update connections.
If the number of clients is equal, compare them one to one.
If there are less clients, see if it is a subset from the previous
client-list.